+++ meja-swing/src/main/java/com/dua3/meja/ui/GraphicsContext.java
@@ -16,6 +16,7 @@
 package com.dua3.meja.ui;
 
 import com.dua3.utility.data.Color;
+import com.dua3.utility.math.geometry.Rectangle2f;
 
 /**
  * The GraphicsContext interface provides methods for drawing various shapes and setting properties related to the graphics.
@@ -30,7 +31,7 @@ public interface GraphicsContext {
      * @param x2 the end point's x-coordinate
      * @param y2 the end point's y-coordinate
      */
-    void drawLine(double x1, double y1, double x2, double y2);
+    void drawLine(float x1, float y1, float x2, float y2);
 
     /**
      * Draws a rectangle with the specified position and size.
@@ -40,7 +41,16 @@ public interface GraphicsContext {
      * @param width  the rectangle's width
      * @param height the rectangle's height
      */
-    void drawRect(double x, double y, double width, double height);
+    void drawRect(float x, float y, float width, float height);
+
+    /**
+     * Draws a rectangle with the same position and size as the given rectangle.
+     *
+     * @param r the rectangle to be drawn
+     */
+    default void drawRect(Rectangle2f r) {
+        drawRect(r.xMin(), r.yMin(), r.width(), r.height());
+    }
 
     /**
      * Fills a rectangle with the specified position and size.
@@ -50,14 +60,23 @@ public interface GraphicsContext {
      * @param width  the rectangle's width
      * @param height the rectangle's height
      */
-    void fillRect(double x, double y, double width, double height);
+    void fillRect(float x, float y, float width, float height);
+
+    /**
+     * Fills a rectangle with the same position and size as the given rectangle.
+     *
+     * @param r the rectangle to be drawn
+     */
+    default void fillRect(Rectangle2f r) {
+        fillRect(r.xMin(), r.yMin(), r.width(), r.height());
+    }
 
     /**
      * Returns the bounds of the current clipping area, specified as a Rectangle object.
      *
      * @return the Rectangle object representing the boundaries of the current clipping area.
      */
-    Rectangle getClipBounds();
+    Rectangle2f getClipBounds();
 
     /**
      * Sets the current color.
@@ -72,7 +91,7 @@ public interface GraphicsContext {
      * @param color the color of the stroke to be set.
      * @param width the width of the stroke to be set.
      */
-    void setStroke(Color color, double width);
+    void setStroke(Color color, float width);
 
     /**
      * Sets the XOR mode for drawing.